Use redux-devtools by hand
I was reading up on the redux dev tools and the really cool part about them is that they aren't redux specific. All you have to do is send 'states' to the dev tools and they'll be visible to be manipulated via the tool. We can then 'subscribe' to the dev tools connection for messages and easily implement the ability to 'time travel' to a specific state.